+++ odoo-autodiscover.py
@@ -3,14 +3,20 @@
 # © 2015 ACSONE SA/NV
 # License LGPLv3 (http://www.gnu.org/licenses/lgpl-3.0-standalone.html)
 
-# this imports the odoo.py script (not completely clean, but there is no other
-# way due to the way the upstream script is packaged)
-import odoo
+import imp
+import sys
+from distutils.spawn import find_executable
 
 from odoo_autodiscover import monkey
 
 
 def main():
+    odoo_py = find_executable('odoo.py')
+    if not odoo_py:
+        sys.stderr.write('The odoo.py executable could not be found, '
+                         'therefore odoo-autodiscover.py can not start.')
+        sys.exit(1)
+    odoo = imp.load_source('odoo', odoo_py)
     monkey.patch()
     odoo.main()