+++ CHANGES.rst
@@ -3,7 +3,9 @@ Changes
 
 1.0.1 (unreleased)
 ------------------
-- 
+- [FIX] odoo-autodiscover.py: more reliable way to discover and import
+  the official odoo.py script, so it will now work when Odoo is installed
+  from the deb package
 
 1.0.0 (2015-12-28)
 ------------------